requirements.txt
requests
pandas
pyjwt
awscli
boto3
pyarrow
pyodbc
fastapi
pydantic
azure-storage-blob
azure-identity
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[argon2-cffi]
argon2-cffi
python-multipart